created Customer Profile View

+++ Capstone/Models/Customer/CustomerProfile.cs
@@ -21,7 +21,7 @@ namespace Capstone.Models
         [Display(Name = "Current Job Title:")]
         public string CustomerJobTitle { get; set; }
         [Display(Name = "What is your expected annual income from this position: $_____")]
-        public int CustomerExpectedAnnualIncome { get; set; }
+        public double CustomerExpectedAnnualIncome { get; set; }
         [Display(Name = "Check the box if you have a second job.")]
         public bool SecondaryJob { get; set; }
         [Display(Name = "What is your secondary job?")] //only show if the above box is checked
@@ -31,14 +31,14 @@ namespace Capstone.Models
         [Display(Name = "What are your other jobs?")] //only show if above box is checked
         public string OtherJobs { get; set; }
         [Display(Name = "What do you expect to earn from jobs other than your primary job? $_____")] //only show if SecondaryJob == true
-        public int AdditionalEarnings1 { get; set; }
+        public double AdditionalEarnings1 { get; set; }
         [Display(Name = "Check the box if you expect to bring in money by means other than working a job.")]
         public bool OtherMeansOfIncome1 { get; set; }
         //if above box is checked
         [Display(Name = "How? Eg. flipping houses, babsitting, collecting rent, receiving alimony check, et cetera")]
         public string SideGig1 { get; set; }
         [Display(Name = "How much do you expect this to bring in this year? $_____")] //if above box is checked
-        public int AdditionalRevenue1 { get; set; } 
+        public double AdditionalRevenue1 { get; set; } 
         //if additionalrevenue1 == true
         [Display(Name = "Do you have other additional means of income?")]
         public bool OtherMeansOfIncome2 { get; set; } 
@@ -46,6 +46,6 @@ namespace Capstone.Models
         [Display(Name = "How? List all other additional means of income. Eg. flipping houses, babsitting, collecting rent, receiving alimony check, et cetera")]
         public string SideGig2 { get; set; } 
         [Display(Name = "How much do you expect this/these to bring in this year? $____")] //if othermeansofincome == true
-        public int AdditionalRevenue2 {get; set;}
+        public double AdditionalRevenue2 {get; set;}
     } 
 }
